feat: Parte 1.2 da AP1

+++ src/domain/Price.java
@@ -1,20 +1,19 @@
 package domain;
 
-import java.math.BigDecimal;
 import java.util.Date;
 import java.util.UUID;
 
 public class Price implements EntityInterface {
     private UUID uuid;
-    private BigDecimal price;
+    private Float price;
     private Date date;
 
-    public Price(Date date, BigDecimal price) {
+    public Price(Date date, Float price) {
         this.price = price;
         this.date = date;
     }
 
-    public Price(BigDecimal price, UUID uuid, Date date) {
+    public Price(Float price, UUID uuid, Date date) {
         this.price = price;
         this.uuid = uuid;
         this.date = date;
